backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
playwright==1.48.*
playwright-stealth==1.0.*
rapidfuzz==3.10.*
pydantic==2.9.*
python-dotenv==1.0.*
pytest==8.3.*
pytest-asyncio==0.24.*
httpx==0.27.*
